LayoutTests/streams/readable-stream-byob-request-worker-expected.txt


FAIL By default, byobRequest should be undefined assert_equals: by default byobRequest should be undefined expected (undefined) undefined but got (object) null
FAIL byobRequest.view length should be equal to autoAllocateChunkSize Default reader on byte stream is not yet supported
FAIL Calling respond() with a this object different from ReadableStreamBYOBRequest should throw a TypeError Default reader on byte stream is not yet supported
FAIL Calling respond() with a negative bytesWritten value should throw a RangeError Default reader on byte stream is not yet supported
FAIL Calling respond() with a bytesWritten value which is not a number should throw a RangeError Default reader on byte stream is not yet supported
FAIL Calling respond() with a positive infinity bytesWritten value should throw a RangeError Default reader on byte stream is not yet supported
FAIL Calling respond() with a bytesWritten value different from 0 when stream is closed should throw a TypeError Default reader on byte stream is not yet supported
FAIL Calling respond() with a bytesWritten value of 0 when stream is closed should succeed Default reader on byte stream is not yet supported
FAIL Calling respond() with a bytesWritten value greater than autoAllocateChunkSize should fail Default reader on byte stream is not yet supported
FAIL Calling respond() with a bytesWritten value lower than autoAllocateChunkSize should succeed Default reader on byte stream is not yet supported
FAIL Calling respondWithNewView() with a this object different from ReadableStreamBYOBRequest should throw a TypeError Default reader on byte stream is not yet supported
FAIL Calling respondWithNewView() with an argument that is not an object should throw a TypeError Default reader on byte stream is not yet supported
FAIL Calling respondWithNewView() with an argument that is not an ArrayBufferView should throw a TypeError Default reader on byte stream is not yet supported
FAIL When using autoAllocateChunkSize, calling respondWithNewView() should succeed if view.byteLength is equal to autoAllocateChunkSize Default reader on byte stream is not yet supported
FAIL When using autoAllocateChunkSize, calling respondWithNewView() should throw a RangeError if view.byteOffset is different from 0 Default reader on byte stream is not yet supported
FAIL When using autoAllocateChunkSize, calling respondWithNewView() should throw a RangeError if view.byteLength is different from autoAllocateChunkSize Default reader on byte stream is not yet supported
